requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-filter>=24.2
django-taggit>=5.0
whitenoise>=6.6
django-redis>=5.4
psycopg2-binary>=2.9
sentry-sdk>=2.3
docxtpl>=0.16
docx2pdf>=0.1.8
